feature/cf cli binary (#57)
Improve performance of CF cli install

The Homebrew CF cli install method takes anywhere from 30s to 1m, while
this binary install method should take less than 5s. We want to get to
the important part of the action we're doing as quickly as possible, so
not wasting time installing tools helps with that.