feature/automate test deployments (#63)
* Remove route info from vars defaults

* Trigger unit test workflow on PRs only

* Change workflow trigger for deployment (run unit tests)

Change the workflow triggers to be directly watching for pushes and tags
on the main branch to allow for more granular workflow trigger control
(watching for tags, specifically). This change also means that
unit-tests will need to be run as part of the deploy workflow to prevent
unneccesary application failures.